adding, removing product in basket

+++ src/store/basket/basket.slice.ts
@@ -5,13 +5,15 @@ import {WritableDraft} from "immer";
 const LS_BSK_KEY = 'bsk'
 
 interface BasketState {
-  baskets: IProduct[]
-  isShowBasket: boolean
+  baskets: IProduct[],
+  totalAmount: number,
+  totalPrice: number,
 }
 
 const initialState: BasketState = {
   baskets: JSON.parse(localStorage.getItem(LS_BSK_KEY) ?? '[]'),
-  isShowBasket: false
+  totalAmount: 0,
+  totalPrice: 0,
 }
 
 export const basketSlice = createSlice({
@@ -19,15 +21,45 @@ export const basketSlice = createSlice({
   initialState,
   reducers: {
     addBasket(state: WritableDraft<BasketState>, action: PayloadAction<IProduct>) {
-      state.baskets.push(action.payload)
+      const isExist = state.baskets.some(product => product.id === action.payload.id)
+      if (isExist) return
+
+      let product: IProduct = {...action.payload}
+      product = {...action.payload, amount: 1}
+      console.log(product)
+
+      state.baskets.push(product)
       localStorage.setItem(LS_BSK_KEY, JSON.stringify(state.baskets))
     },
-    removeBasket(state, action: PayloadAction<IProduct>) {
-      state.baskets = state.baskets.filter(f => f !== action.payload)
+    increaseCountOfProduct(state, action: PayloadAction<IProduct>) {
+      state.baskets.forEach((item: any) => {
+        if (item.id === action.payload.id) {
+          item.amount++
+        }
+      })
+
       localStorage.setItem(LS_BSK_KEY, JSON.stringify(state.baskets))
     },
-    toggleBasket(state) {
-      state.isShowBasket = !state.isShowBasket
+    decreaseCountOfProduct(state, action: PayloadAction<IProduct>) {
+
+      state.baskets.forEach((item: any) => {
+        if (item.id === action.payload.id) {
+          if (item.amount === 1) {
+            state.baskets = state.baskets.filter(product => product.id !== action.payload.id)
+          } else {
+            item.amount--
+          }
+        }
+      })
+
+      localStorage.setItem(LS_BSK_KEY, JSON.stringify(state.baskets))
+    },
+    removeAllProductInBasket(state) {
+      const isConfirm = confirm('Do you really want to do this?')
+      if (!isConfirm) return
+
+      state.baskets = []
+      localStorage.setItem(LS_BSK_KEY, JSON.stringify(state.baskets))
     },
   }
 })
